start index helper refactor
